// src/app/shared/enums/dashboard-panal/dashboard-panal.enum.ts
export enum DashboardPanalEnums{
    EDUCATIONAL_SETTING = 'EDUCATIONAL_SETTING',
    REPORTS_MANAGEMENT = 'REPORTS_MANAGEMENT',
    SCHOOLS_AND_STUDENTS = 'SCHOOLS_AND_STUDENTS',
    MANAGAR_TOOLS = 'MANAGAR_TOOLS',
    PEFORMANCE_MANAGMENT = 'PEFORMANCE_MANAGMENT',
    SCHOOL_STUDENTS = 'SCHOOL_STUDENTS',
    COMMUNICATION_MANAGMENT ='COMMUNICATION_MANAGMENT',
    SCHOOL_INFO = 'SCHOOL_INFO',
    School_Management = 'School_Management',
    Grades_Divisions_Management = 'Grades_Divisions_Management',
    SchoolEmployee_Management = 'SchoolEmployee_Management',
    Student_Management = 'Student_Management',
    School_PerformanceManagent = 'School_PerformanceManagent'
  }